services/order-services/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
python-dotenv
httpx
alembic